Fixed cart update from catalog item

+++ src/hooks/useUpdateCart.js
@@ -8,36 +8,45 @@ const useUpdateCart = (url, headers, item, index = undefined) => {
     const { dispatch, cartItems } = useCart();
 
     const updateCart = (quantity) => {
-        // Clone cartItems array to trigger cartItems state change @ CartContext
-        const items = [...cartItems];
+        // Cart is not empty
+        if (cartItems.length) {
+            // Clone cartItems array
+            // Note: empty cart is not iteralbe [can't be destructured]
+            const items = [...cartItems];
+            // // If quantity is zero
+            if (quantity === 0) {
+                items.splice(index, 1);
 
-        // If quantity is zero
-        if (quantity === 0) {
-            items.splice(index, 1);
-            // Remove item form cart @ CartContext
-            dispatch({
-                type: 'delete-item',
-                payload: items
-            })
-        }
-        else {
-            // If index is defined [product already in cart]
-            if (index !== undefined) {
-                console.log(index)
-                // Update product quantity
-                items[index].quantity = quantity;
+                dispatch({
+                    type: 'delete-item',
+                    payload: {
+                        cartItems: items,
+                    }
+                })
             }
             else {
-                // Else add item in cart [new item]
-                items.push({ selected: item, quantity: quantity });
-            }
-            // Update cart reducer @ CartContext
-            dispatch({
-                type: 'success',
-                payload: {
-                    cartItems: items
+                // If index is defined [product already in cart]
+                if (index !== undefined) {
+                // Update product quantity
+                items[index].quantity = quantity;
                 }
-            });
+                else {
+                    // Else add item in cart [new item]
+                    items.push({ selected: item, quantity: quantity });
+                }
+                // Update cart reducer @ CartContext
+                dispatch({
+                    type: 'set-cart',
+                    payload: items
+                });
+            }
+        }
+        // Cart is empty
+        else {
+            dispatch({
+                type: 'set-cart',
+                payload: [{ selected: item, quantity: quantity }]
+            })
         }
     }
 
@@ -45,20 +54,20 @@ const useUpdateCart = (url, headers, item, index = undefined) => {
     // Receives an integer as argument [quantitySelected]
     // @ QuantitySelector component
     // @ CollectionItem component [delete button]
-    const updateCartHandler = (quantitySelected) => {
-        // Parse string value
+    const updateCartHandler = (quantity) => {
 
-            // Set loader
-            setLoading(true);
-            axios.get(`${api.serverURL}${url}${quantitySelected}`, headers)
+        setLoading(true);
+
+        axios.get(`${api.serverURL}${url}${quantity}`, headers)
                 .then(res => {
                     // If quantity selected is available
                     if (res.status === 200 && res.data.isAvailable) {
                         // Update cart quantities
-                        updateCart(quantitySelected);
+                        updateCart(quantity);
                     }
                 })
                 .catch((error) => {
+                    // setQuantitySelected(0);
                     dispatch({
                         type: 'error',
                         payload: {
